+++ spec/aws/active_job/sqs/poller_spec.rb
@@ -47,10 +47,10 @@ module Aws
             allow(poller).to receive(:poll) # no-op the poll
             poller.run
 
-            options = poller.instance_variable_get(:@options)
-            expect(options[:max_messages]).to eq 5 # from test app config file
-            expect(options[:visibility_timeout]).to eq 360 # from argv
-            expect(options[:shutdown_timeout]).to eq 15 # from defaults
+            config = Aws::ActiveJob::SQS.config
+            expect(config.max_messages_for(:default)).to eq 2 # from test app config file
+            expect(config.visibility_timeout_for(:default)).to eq 360 # from argv
+            expect(config.shutdown_timeout).to eq 15 # from defaults
           end
 
           it 'polls the configured queue' do
@@ -71,7 +71,7 @@ module Aws
             expect(queue_poller).to receive(:poll).with(
               {
                 skip_delete: true,
-                max_number_of_messages: 5,
+                max_number_of_messages: 2, # from queue config in app config file
                 visibility_timeout: 360
               }
             )
@@ -82,7 +82,7 @@ module Aws
           it 'sets max_number_of_messages to 1 for fifo queues' do
             allow(poller).to receive(:boot_rails) # no-op the boot
 
-            allow(Aws::ActiveJob::SQS.config).to receive(:queue_url_for).and_return('https://queue-url.fifo')
+            allow(Aws::ActiveJob::SQS.config).to receive(:url_for).and_return('https://queue-url.fifo')
             expect(Aws::SQS::QueuePoller).to receive(:new).and_return(queue_poller)
 
             expect(queue_poller).to receive(:poll).with(